+++ crates/node/src/service.rs
@@ -12,6 +12,7 @@ use futures::prelude::*;
 use madara_runtime::opaque::Block;
 use madara_runtime::{self, Hash, RuntimeApi, SealingMode, StarknetHasher};
 use mc_block_proposer::ProposerFactory;
+use mc_commitment_state_diff::{log_commitment_state_diff, CommitmentStateDiffWorker};
 use mc_data_availability::avail::config::AvailConfig;
 use mc_data_availability::avail::AvailClient;
 use mc_data_availability::celestia::config::CelestiaConfig;
@@ -384,6 +385,21 @@ pub fn new_full(
             0,
         )
         .for_each(|()| future::ready(())),
+    );
+
+    let (commitment_state_diff_tx, commitment_state_diff_rx) = mpsc::channel(5);
+
+    task_manager.spawn_essential_handle().spawn(
+        "commitment-state-diff",
+        Some("madara"),
+        CommitmentStateDiffWorker::<_, _, StarknetHasher>::new(client.clone(), commitment_state_diff_tx)
+            .for_each(|()| future::ready(())),
+    );
+
+    task_manager.spawn_essential_handle().spawn(
+        "commitment-state-logger",
+        Some("madara"),
+        log_commitment_state_diff(commitment_state_diff_rx),
     );
 
     // initialize data availability worker