+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 ## Next release
 
 - test: add transaction pool logic unit tests
+- feat(client): spawn a task that listen to storage changes and build the
+  resulting commiment state diff for each block
+- dev(StarknetRPC): log error received from node before mapping to
+  InternalServerError
 - fix: change 'nonce too high' to log in debug instead of info
 - chore: update deps, vm ressource fee cost are now FixedU128, and stored in an
   hashmap
@@ -39,6 +43,8 @@
   method being used)
 - refactor(sealing): how the sealing mode is passed into runtime
 - feat(sealing): finalization for instant sealing
+- test(starknet-js-test): run basic starknetjs compatibility tests again the
+  madara node
 - feat(cache-option): add an option to enable aggressive caching in command-line
   parameters